requirements.txt
cloudscraper
beautifulsoup4
requests
google-generativeai